refactor export view components, update heading text
