+++ packages/raga-app/package.json
@@ -49,6 +49,7 @@
     "milliparsec": "^5.0.1",
     "motion": "^12.4.1",
     "node-taglib-sharp": "^6.0.1",
+    "pluralize": "^8.0.0",
     "radash": "^12.1.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
@@ -73,6 +74,7 @@
     "@types/app-root-dir": "^0.1.4",
     "@types/chroma-js": "^3.1.1",
     "@types/node": "^22.13.1",
+    "@types/pluralize": "^0.0.33",
     "@types/react": "^19.0.8",
     "@types/react-devtools": "^3.6.2",
     "@types/react-dom": "^19.0.3",
